+++ docs/inference.md
@@ -1,3 +1,5 @@
 Inference
 
 Llava-omni - https://github.com/ictnlp/LLaMA-Omni
+
+https://build.nvidia.com/meta/llama-3.2-11b-vision-instruct
